drm/msm/sde: add new clock/bandwidth capabilities
Add new capability fields to enable clock and bandwidth
calculation in user mode. These new fields expose addtional
parameters, specific to a particular chip set and platform,
obtained from catalog and/or device tree.

CRs-Fixed: 2009714

+++ drivers/gpu/drm/msm/sde/sde_kms.h
@@ -325,11 +325,11 @@ void sde_kms_info_reset(struct sde_kms_info *info);
  * sde_kms_info_add_keyint - add integer value to 'sde_kms_info'
  * @info: Pointer to sde_kms_info structure
  * @key: Pointer to key string
- * @value: Signed 32-bit integer value
+ * @value: Signed 64-bit integer value
  */
 void sde_kms_info_add_keyint(struct sde_kms_info *info,
 		const char *key,
-		int32_t value);
+		int64_t value);
 
 /**
  * sde_kms_info_add_keystr - add string value to 'sde_kms_info'
